Check in example of crash using only safe Rust I ran into.
There seems to be a problem with the println! macro combined with
iterators. Passing <iterator item>.<&'static str member> directly
to println! causes a crash, but first binding
<iterator item>.<&'static str member> to a variable with let then
passing that variable to println! works okay.

This could be bug compiler or library bug. Happens in stable Rust 1.24.1
in both debug and release builds. Get unexpected segfault or panic
with error about invalid unicode.